extract imperative handle to separate file
